requirements.txt
pyupbit==0.2.33
pandas>=2.2.3
numpy>=2.1.0
schedule>=1.2.1
requests>=2.31.0
python-dotenv>=1.0.1
